+++ priv/repo/migrations/20170526125941_add_status_to_idea.exs
@@ -1,0 +1,9 @@
+defmodule Icta.Repo.Migrations.AddStatusToIdea do
+  use Ecto.Migration
+
+  def change do
+    alter table(:ideas) do
+      add :status, :string, default: "new", null: false
+    end
+  end
+end