+++ web/static/js/actions/ideas.js
@@ -5,21 +5,25 @@ import { I18n } from 'react-redux-i18n';
 let socket = configureChannel();
 let channel = socket.channel('icta', {});
 
-export const FETCH_IDEAS_REQUEST = 'FETCH_IDEAS_REQUEST';
-export const FETCH_IDEAS_SUCCESS = 'FETCH_IDEAS_SUCCESS';
-export const FETCH_IDEAS_FAILURE = 'FETCH_IDEAS_FAILURE';
+export const SHOW_IDEAS_REQUEST = 'SHOW_IDEAS_REQUEST';
+export const SHOW_IDEAS_SUCCESS = 'SHOW_IDEAS_SUCCESS';
+export const SHOW_IDEAS_FAILURE = 'SHOW_IDEAS_FAILURE';
 
 export const ADD_IDEA_REQUEST = 'ADD_IDEA_REQUEST';
 export const ADD_IDEA_SUCCESS = 'ADD_IDEA_SUCCESS';
 export const ADD_IDEA_FAILURE = 'ADD_IDEA_FAILURE';
 
+export const EDIT_IDEA_REQUEST = 'EDIT_IDEA_REQUEST';
+export const EDIT_IDEA_SUCCESS = 'EDIT_IDEA_SUCCESS';
+export const EDIT_IDEA_FAILURE = 'EDIT_IDEA_FAILURE';
+
 export const VOTE_REQUEST = 'VOTE_REQUEST';
 export const VOTE_SUCCESS = 'VOTE_SUCCESS';
 export const VOTE_FAILURE = 'VOTE_FAILURE';
 
-export const SHOW_IDEA_SUCCESS = 'SHOW_IDEA_SUCCESS';
-export const SHOW_IDEA_REQUEST = 'SHOW_IDEA_REQUEST';
-export const SHOW_IDEA_FAILURE = 'SHOW_IDEA_FAILURE';
+export const FETCH_IDEA_SUCCESS = 'FETCH_IDEA_SUCCESS';
+export const FETCH_IDEA_REQUEST = 'FETCH_IDEA_REQUEST';
+export const FETCH_IDEA_FAILURE = 'FETCH_IDEA_FAILURE';
 
 export const DELETE_IDEA_SUCCESS = 'DELETE_IDEA_SUCCESS';
 export const DELETE_IDEA_REQUEST = 'DELETE_IDEA_REQUEST';
@@ -34,6 +38,7 @@ export const ADD_COMMENT_SUCCESS = 'ADD_COMMENT_SUCCESS';
 export const ADD_COMMENT_FAILURE = 'ADD_COMMENT_FAILURE';
 
 export const NEW_IDEA_RECEIVED = 'NEW_IDEA_RECEIVED';
+export const EDIT_IDEA_RECEIVED = 'EDIT_IDEA_RECEIVED';
 export const CHANGE_MY_VOTE = 'CHANGE_MY_VOTE';
 
 const userInfoRequest = () => ({ type: USER_INFO_REQUEST });
@@ -44,17 +49,21 @@ const voteRequest = () => ({ type: VOTE_REQUEST });
 const voteSuccess = (idea) => ({ type: VOTE_SUCCESS, idea });
 const voteFailure = (error) => ({ type: VOTE_FAILURE, error });
 
-const fetchIdeasRequest = () => ({ type: FETCH_IDEAS_REQUEST });
-const fetchIdeasSuccess = ideas => ({ type: FETCH_IDEAS_SUCCESS, ideas });
-const fetchIdeasFailure = error => ({ type: FETCH_IDEAS_FAILURE, ideas });
+const showIdeasRequest = () => ({ type: SHOW_IDEAS_REQUEST });
+const showIdeasSuccess = ideas => ({ type: SHOW_IDEAS_SUCCESS, ideas });
+const showIdeasFailure = error => ({ type: SHOW_IDEAS_FAILURE, ideas });
 
 const addIdeaRequest = title => ({ type: ADD_IDEA_REQUEST, title, body });
 const addIdeaSuccess = idea => ({ type: ADD_IDEA_SUCCESS });
 const addIdeaFailure = (title, error) => ({ type: ADD_IDEA_FAILURE, title, error });
 
-const showIdeaRequest = () => ({ type: SHOW_IDEA_REQUEST });
-const showIdeaSuccess = (idea) => ({ type: SHOW_IDEA_SUCCESS, idea });
-const showIdeaFailure = (error) => ({ type: SHOW_IDEA_FAILURE, error });
+const editIdeaRequest = () => ({ type: EDIT_IDEA_REQUEST });
+const editIdeaSuccess = () => ({ type: EDIT_IDEA_SUCCESS });
+const editIdeaFailure = (error) => ({ type: EDIT_IDEA_FAILURE, error });
+
+const fetchIdeaRequest = () => ({ type: FETCH_IDEA_REQUEST });
+const fetchIdeaSuccess = (idea) => ({ type: FETCH_IDEA_SUCCESS, idea });
+const fetchIdeaFailure = (error) => ({ type: FETCH_IDEA_FAILURE, error });
 
 const addCommentRequest = () => ({ type: ADD_COMMENT_REQUEST });
 const addCommentSuccess = () => ({ type: ADD_COMMENT_SUCCESS });
@@ -65,6 +74,7 @@ const deleteIdeaSuccess = (ideaId) => ({ type: DELETE_IDEA_SUCCESS, ideaId });
 const deleteIdeaFailure = (error) => ({ type: DELETE_IDEA_FAILURE, error });
 
 const newIdeaReceived = idea => ({ type: NEW_IDEA_RECEIVED, idea });
+const editIdeaReceived = idea => ({ type: EDIT_IDEA_RECEIVED, idea });
 
 const changeMyVote = (ideaId, myVote) => ({ type: CHANGE_MY_VOTE, ideaId, myVote });
 
@@ -123,7 +133,7 @@ export const addComment = (ideaId, body) => (
     channel.push('idea:comment:new', payload)
       .receive('ok', response => {
         dispatch(addCommentSuccess());
-        dispatch(showIdea(ideaId));
+        dispatch(fetchIdea(ideaId));
       })
       .receive('error', error => {
         dispatch(addCommentFailure(error));
@@ -131,9 +141,9 @@ export const addComment = (ideaId, body) => (
   }
 );
 
-export const showIdea = (ideaId) => (
+export const fetchIdea = (ideaId) => (
   dispatch => {
-    dispatch(showIdeaRequest());
+    dispatch(fetchIdeaRequest());
 
     const payload = { idea_id: ideaId }
 
@@ -142,17 +152,17 @@ export const showIdea = (ideaId) => (
 
         const comments = [ ...response.comments ]
         const idea = { ...response.idea, comments };
-        dispatch(showIdeaSuccess(idea));
+        dispatch(fetchIdeaSuccess(idea));
       })
       .receive('error', error => {
-        dispatch(showIdeaFailure(error));
+        dispatch(fetchIdeaFailure(error));
       });
   }
 );
 
-export const fetchIdeas = () => (
+export const showIdeas = () => (
   dispatch => {
-    dispatch(fetchIdeasRequest());
+    dispatch(showIdeasRequest());
 
     channel.join()
       .receive('ok', messages => {
@@ -164,14 +174,18 @@ export const fetchIdeas = () => (
 
     channel.push('idea:get_all', null)
       .receive('ok', messages => {
-        dispatch(fetchIdeasSuccess(messages.ideas));
+        dispatch(showIdeasSuccess(messages.ideas));
       })
       .receive('error', reason => {
-        dispatch(fetchIdeasFailure(reason));
+        dispatch(showIdeasFailure(reason));
       });
 
     channel.on('idea:new', msg => {
       dispatch(newIdeaReceived(msg));
+    });
+
+    channel.on('idea:edit', msg => {
+      dispatch(editIdeaReceived(msg));
     });
 
     channel.on('vote:new', msg => {
@@ -199,6 +213,30 @@ export const addIdea = (title, body, history) => (
       })
       .receive('error', error => {
         dispatch(addIdeaFailure(title, error));
+      });
+  }
+);
+
+export const editIdea = (ideaId, attributes, history) => (
+  dispatch => {
+    dispatch(editIdeaRequest());
+
+    const payload = { idea_id: ideaId, attributes }
+
+    channel.push('idea:edit', payload)
+      .receive('ok', response => {
+        dispatch(editIdeaSuccess());
+
+        dispatch(Notifications.success({
+          title: I18n.t('notifications.edit_idea_success.title'),
+          message: I18n.t('notifications.edit_idea_success.message'),
+        }));
+
+        history.push('/');
+      })
+
+      .receive('error', error => {
+        dispatch(editIdeaFailure(error));
       });
   }
 );