+++ web/static/js/components/IdeasListItem.jsx
@@ -1,31 +1,43 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom'
-import { vote, deleteIdea, showIdea } from '../actions/ideas';
+import { vote, deleteIdea, fetchIdea } from '../actions/ideas';
+import { Translate } from 'react-redux-i18n';
 
 const IdeasListItem = withRouter(({idea, deleteIdea, voteUp, voteDown, ideaClick, history, userId}) => (
   <li className="list-group-item">
     <div className="row">
-      <div className="col-sm-2 col-xs-3 thumbs-group text-center">
-        <div className="thumbs">
-          <a onClick={() => {voteUp(idea.id)}}>
-            <i className={`fa fa-thumbs-up fa-2x ${idea.my_vote === true ? 'active' : ''}`}></i>
-          </a><br />
-          <span className="vote-count">{idea.up}</span><br />
-        </div>
-        <div className="thumbs">
-          <a onClick={() => {voteDown(idea.id)}}>
-            <i className={`fa fa-thumbs-down fa-2x ${idea.my_vote === false ? 'active' : ''}`}></i>
-          </a><br />
-          <span className="vote-count">{idea.down}</span><br />
+      <div className="col-xs-8 ">
+        <div className="row">
+          <div className="col-xs-5 col-md-5 col-lg-4">
+            <div className="thumbs-group text-center" >
+              <div className="thumbs">
+                <a onClick={() => {voteUp(idea.id)}}>
+                  <i className={`fa fa-thumbs-up fa-2x ${idea.my_vote === true ? 'active' : ''}`}></i>
+                </a><br />
+                <span className="vote-count">{idea.up}</span><br />
+              </div>
+              <div className="thumbs">
+                <a onClick={() => {voteDown(idea.id)}}>
+                  <i className={`fa fa-thumbs-down fa-2x ${idea.my_vote === false ? 'active' : ''}`}></i>
+                </a><br />
+                <span className="vote-count">{idea.down}</span><br />
+              </div>
+            </div>
+            <div className="status-badge" onClick={() => {ideaClick(idea, history)}}>
+              <h4><span className={`label label-status-${idea.status}`}><Translate value={`idea.statuses.${idea.status}`} dangerousHTML /></span></h4>
+            </div>
+          </div>
+          <div className="col-xs-7 col-md-7 col-lg-8 idea-text" onClick={() => {ideaClick(idea, history)}}>
+            <h4>{idea.title} </h4>
+            <p className="author">
+              <img className="small-profile-image" src={idea.author.image_url} />
+              &nbsp; <strong>{idea.author.name}</strong>
+            </p>
+          </div>
         </div>
       </div>
-      <div className="col-sm-7 col-xs-6" onClick={() => {ideaClick(idea, history)}}>
-        <h4>{idea.title}</h4>
-        <p className="author">
-          <img className="small-profile-image" src={idea.author.image_url} />
-          &nbsp; <strong>{idea.author.name}</strong>
-        </p>
+      <div className="col-xs-1 idea-status text-center">
       </div>
       <div className="col-xs-2 idea-comments-count text-right" onClick={() => {ideaClick(idea, history)}}>
         <i className="fa fa-comments-o fa-2x"></i> <span>{idea.comments_count}</span>