+++ web/static/js/i18n/en-US.js
@@ -10,6 +10,12 @@ export default {
       new: 'Comment',
       title: 'Comments',
     },
+    statuses: {
+      new: 'New',
+      planned: 'Planned',
+      in_progress: 'In progress',
+      done: 'Finished',
+    },
     title: 'Title',
     body: 'Body',
     new: 'Create idea',
@@ -18,6 +24,10 @@ export default {
     new_idea_success: {
       title: 'Idea created!',
       message: 'Idea created successfully.',
+    },
+    edit_idea_success: {
+      title: 'Idea updated!',
+      message: 'Idea updated successfully.',
     },
   },
   datetime: 'MM/DD/YYYY, h:mm',