+++ web/static/js/i18n/pt-BR.js
@@ -10,6 +10,12 @@ export default {
       new: 'Comentar',
       title: 'Comentários',
     },
+    statuses: {
+      new: 'Nova',
+      planned: 'Planejada',
+      in_progress: 'Em progresso',
+      done: 'Finalizada',
+    },
     title: 'Título',
     body: 'Conteúdo',
     new: 'Criar ideia',
@@ -18,6 +24,10 @@ export default {
     new_idea_success: {
       title: 'Ideia criada!',
       message: 'Ideia criada com sucesso.',
+    },
+    edit_idea_success: {
+      title: 'Ideia alterada!',
+      message: 'Ideia alterada com sucesso.',
     },
   },
   datetime: 'DD/MM/YYYY [as] HH:mm',